+++ contrib/platform/src/com/sun/jna/platform/win32/WinPerf.java
@@ -23,8 +23,8 @@
  */
 package com.sun.jna.platform.win32;
 
-import java.util.List;
-
+import com.sun.jna.NativeLong;
+import com.sun.jna.Pointer;
 import com.sun.jna.Structure;
 import com.sun.jna.Structure.FieldOrder;
 import com.sun.jna.platform.win32.WinBase.SYSTEMTIME;
@@ -60,6 +60,15 @@ public interface WinPerf {
         public LARGE_INTEGER PerfTime100nSec = new LARGE_INTEGER();
         public int           SystemNameLength;
         public int           SystemNameOffset;
+
+        public PERF_DATA_BLOCK() {
+            super();
+        }
+
+        public PERF_DATA_BLOCK(Pointer p) {
+            super(p);
+            read();
+        }
     };
 
     /**
@@ -75,6 +84,106 @@ public interface WinPerf {
         public int UniqueID;
         public int NameOffset;
         public int NameLength;
+
+        public PERF_INSTANCE_DEFINITION() {
+            super();
+        }
+
+        public PERF_INSTANCE_DEFINITION(Pointer p) {
+            super(p);
+            read();
+        }
+    }
+
+    /**
+     * Describes object-specific performance information, for example, the
+     * number of instances of the object and the number of counters that the
+     * object defines.
+     * 
+     * @see <A HREF=
+     *      "https://msdn.microsoft.com/en-us/library/windows/desktop/aa373160(v=vs.85).aspx">
+     *      PERF_OBJECT_TYPE</A>
+     */
+    @FieldOrder({ "TotalByteLength", "DefinitionLength", "HeaderLength", "ObjectNameTitleIndex", "ObjectNameTitle",
+            "ObjectHelpTitleIndex", "ObjectHelpTitle", "DetailLevel", "NumCounters", "DefaultCounter", "NumInstances",
+            "CodePage", "PerfTime", "PerfFreq" })
+    public class PERF_OBJECT_TYPE extends Structure {
+        public int TotalByteLength;
+        public int DefinitionLength;
+        public int HeaderLength;
+        public int ObjectNameTitleIndex;
+        public int ObjectNameTitle; // always 32 bit
+        public int ObjectHelpTitleIndex;
+        public int ObjectHelpTitle; // always 32 bit
+        public int DetailLevel;
+        public int NumCounters;
+        public int DefaultCounter;
+        public int NumInstances;
+        public int CodePage;
+        public LARGE_INTEGER PerfTime;
+        public LARGE_INTEGER PerfFreq;
+
+        public PERF_OBJECT_TYPE() {
+            super();
+        }
+
+        public PERF_OBJECT_TYPE(Pointer p) {
+            super(p);
+            read();
+        }
+    }
+
+    /**
+     * Describes a performance counter.
+     * 
+     * @see <A HREF=
+     *      "https://msdn.microsoft.com/en-us/library/windows/desktop/aa373150(v=vs.85).aspx">
+     *      PERF_COUNTER_DEFINITION</A>
+     */
+    @FieldOrder({ "ByteLength", "CounterNameTitleIndex", "CounterNameTitle", "CounterHelpTitleIndex",
+            "CounterHelpTitle", "DefaultScale", "DetailLevel", "CounterType", "CounterSize", "CounterOffset" })
+    public class PERF_COUNTER_DEFINITION extends Structure {
+        public int ByteLength;
+        public int CounterNameTitleIndex;
+        public int CounterNameTitle; // always 32 bit
+        public int CounterHelpTitleIndex;
+        public int CounterHelpTitle; // always 32 bit
+        public NativeLong DefaultScale;
+        public int DetailLevel;
+        public int CounterType;
+        public int CounterSize;
+        public int CounterOffset;
+
+        public PERF_COUNTER_DEFINITION() {
+            super();
+        }
+
+        public PERF_COUNTER_DEFINITION(Pointer p) {
+            super(p);
+            read();
+        }
+    }
+
+    /**
+     * Describes the block of memory that contains the raw performance counter
+     * data for an object's counters.
+     * 
+     * @see <A HREF=
+     *      "https://msdn.microsoft.com/en-us/library/windows/desktop/aa373147(v=vs.85).aspx">
+     *      PERF_COUNTER_BLOCK</A>
+     */
+    @FieldOrder({ "ByteLength" })
+    public class PERF_COUNTER_BLOCK extends Structure {
+        public int ByteLength;
+
+        public PERF_COUNTER_BLOCK() {
+            super();
+        }
+
+        public PERF_COUNTER_BLOCK(Pointer p) {
+            super(p);
+            read();
+        }
     }
 
     int PERF_NO_INSTANCES = -1;  // no instances (see NumInstances above)